Windows updates
git-svn-id: https://ravl.svn.sourceforge.net/svnroot/ravl/main/Ravl@5504 7226f50c-d813-0410-99bf-979b288ab488

+++ OS/Time/Date.cc
@@ -206,7 +206,7 @@ namespace RavlN {
 #if RAVL_OS_WIN32
     struct _timeb ltime;
     _ftime(&ltime);
-	return DateC(ltime.time, ltime.millitm * 1000);
+	return DateC((long)ltime.time, (long)ltime.millitm * 1000);
 #else
     throw ExceptionC("DateC::NowUTC(), Not implemented. ");
 #endif
@@ -224,7 +224,7 @@ namespace RavlN {
 #if RAVL_OS_WIN32
     struct _timeb ltime;
     _ftime(&ltime);
-    return DateC(ltime.time, ltime.millitm * 1000);
+    return DateC((long)ltime.time, (long)ltime.millitm * 1000);
 #else
     throw ExceptionC("DateC::NowLocal(), Not implemented. ");
 #endif
